Split admin.ts into modular route files
Reorganizes the monolithic admin.ts into logical modules:
- admin/dashboard.ts - dashboard handler
- admin/auth.ts - login/logout routes
- admin/settings.ts - password and Stripe configuration
- admin/sessions.ts - session management
- admin/events.ts - event CRUD operations
- admin/attendees.ts - attendee management
- admin/utils.ts - shared types and constants
- admin/index.ts - combines all routes
