--- src/routes/admin.ts
@@ -1,587 +1,0 @@
-/**
- * Admin routes - dashboard, events, settings, authentication
- */
-
-import {
-  clearLoginAttempts,
-  createSession,
-  deleteAttendee,
-  deleteEvent,
-  deleteOtherSessions,
-  deleteSession,
-  type EventInput,
-  eventsTable,
-  getAllEvents,
-  getAllSessions,
-  getAttendee,
-  getAttendees,
-  getEventWithCount,
-  hasStripeKey,
-  isLoginRateLimited,
-  recordFailedLogin,
-  updateAdminPassword,
-  updateStripeKey,
-  verifyAdminPassword,
-} from "#lib/db";
-import { validateForm } from "#lib/forms.tsx";
-import {
-  createHandler,
-  defineResource,
-  deleteHandler,
-  updateHandler,
-} from "#lib/rest";
-import type { EventWithCount } from "#lib/types.ts";
-import {
-  adminDashboardPage,
-  adminDeleteAttendeePage,
-  adminDeleteEventPage,
-  adminEventEditPage,
-  adminEventPage,
-  adminLoginPage,
-  adminSessionsPage,
-  adminSettingsPage,
-  changePasswordFields,
-  eventFields,
-  generateAttendeesCsv,
-  loginFields,
-  stripeKeyFields,
-} from "#templates";
-import {
-  createRouter,
-  defineRoutes,
-  type RouteHandlerFn,
-  type RouteParams,
-} from "./router.ts";
-import type { ServerContext } from "./types.ts";
-import {
-  type AuthSession,
-  generateSecureToken,
-  getClientIp,
-  htmlResponse,
-  isAuthenticated,
-  notFoundResponse,
-  parseFormData,
-  redirect,
-  requireAuthForm,
-  requireSessionOr,
-  withAuthForm,
-  withEvent,
-  withSession,
-} from "./utils.ts";
-
-/** Form field definition type */
-type FormFields = Parameters<typeof validateForm>[1];
-
-/** Result of form validation with typed values */
-type ValidatedForm = ReturnType<typeof validateForm> & { valid: true };
-
-/** Auth + form + validation result */
-type AuthValidationResult =
-  | { ok: true; session: AuthSession; validation: ValidatedForm }
-  | { ok: false; response: Response };
-
-/** Require auth + form + validation, with custom error handler */
-const requireAuthValidation = async (
-  request: Request,
-  fields: FormFields,
-  onError?: (
-    session: AuthSession,
-    form: URLSearchParams,
-    error: string,
-  ) => Response | Promise<Response>,
-): Promise<AuthValidationResult> => {
-  const auth = await requireAuthForm(request);
-  if (!auth.ok) return auth;
-
-  const validation = validateForm(auth.form, fields);
-  if (!validation.valid) {
-    const errorResponse = onError
-      ? await onError(auth.session, auth.form, validation.error)
-      : redirect("/admin/");
-    return { ok: false, response: errorResponse };
-  }
-
-  return { ok: true, session: auth.session, validation };
-};
-
-/** Extract event input from validated form */
-const extractEventInput = (values: Record<string, unknown>): EventInput => ({
-  name: values.name as string,
-  description: values.description as string,
-  maxAttendees: values.max_attendees as number,
-  thankYouUrl: values.thank_you_url as string,
-  unitPrice: values.unit_price as number | null,
-  maxQuantity: values.max_quantity as number,
-  webhookUrl: (values.webhook_url as string) || null,
-});
-
-/** Events resource for REST operations */
-const eventsResource = defineResource({
-  table: eventsTable,
-  fields: eventFields,
-  toInput: extractEventInput,
-  nameField: "name",
-  onDelete: (id) => deleteEvent(id as number),
-});
-
-/** Attendee type */
-type Attendee = Awaited<ReturnType<typeof getAttendees>>[number];
-
-/** Handle event with attendees - auth, fetch, then apply handler fn */
-const withEventAttendees = async (
-  request: Request,
-  eventId: number,
-  handler: (event: EventWithCount, attendees: Attendee[]) => Response,
-): Promise<Response> => {
-  if (!(await isAuthenticated(request))) {
-    return redirect("/admin/");
-  }
-  return withEvent(eventId, async (event) =>
-    handler(event, await getAttendees(eventId)),
-  );
-};
-
-/** Login page response helper */
-const loginResponse = (error?: string, status = 200) =>
-  htmlResponse(adminLoginPage(error), status);
-
-/**
- * Handle GET /admin/
- */
-const handleAdminGet = (request: Request): Promise<Response> =>
-  withSession(
-    request,
-    async (session) =>
-      htmlResponse(adminDashboardPage(await getAllEvents(), session.csrfToken)),
-    () => loginResponse(),
-  );
-
-/**
- * Handle POST /admin/login
- */
-const handleAdminLogin = async (
-  request: Request,
-  server?: ServerContext,
-): Promise<Response> => {
-  const clientIp = getClientIp(request, server);
-
-  // Check rate limiting
-  if (await isLoginRateLimited(clientIp)) {
-    return loginResponse(
-      "Too many login attempts. Please try again later.",
-      429,
-    );
-  }
-
-  const form = await parseFormData(request);
-  const validation = validateForm(form, loginFields);
-
-  if (!validation.valid) {
-    return loginResponse(validation.error, 400);
-  }
-
-  const valid = await verifyAdminPassword(validation.values.password as string);
-  if (!valid) {
-    await recordFailedLogin(clientIp);
-    return loginResponse("Invalid credentials", 401);
-  }
-
-  // Clear failed attempts on successful login
-  await clearLoginAttempts(clientIp);
-
-  const token = generateSecureToken();
-  const csrfToken = generateSecureToken();
-  const expires = Date.now() + 24 * 60 * 60 * 1000; // 24 hours
-  await createSession(token, csrfToken, expires);
-
-  return redirect(
-    "/admin/",
-    `__Host-session=${token}; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=86400`,
-  );
-};
-
-/** Cookie to clear admin session */
-const clearSessionCookie =
-  "__Host-session=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0";
-
-/**
- * Handle GET /admin/logout
- */
-const handleAdminLogout = (request: Request): Promise<Response> =>
-  withSession(
-    request,
-    async (session) => {
-      await deleteSession(session.token);
-      return redirect("/admin/", clearSessionCookie);
-    },
-    () => redirect("/admin/", clearSessionCookie),
-  );
-
-/**
- * Handle GET /admin/settings
- */
-const handleAdminSettingsGet = (request: Request): Promise<Response> =>
-  requireSessionOr(request, async (session) =>
-    htmlResponse(adminSettingsPage(session.csrfToken, await hasStripeKey())),
-  );
-
-/**
- * Validate change password form data
- */
-type ChangePasswordValidation =
-  | { valid: true; currentPassword: string; newPassword: string }
-  | { valid: false; error: string };
-
-const validateChangePasswordForm = (
-  form: URLSearchParams,
-): ChangePasswordValidation => {
-  const validation = validateForm(form, changePasswordFields);
-  if (!validation.valid) {
-    return validation;
-  }
-
-  const { values } = validation;
-  const currentPassword = values.current_password as string;
-  const newPassword = values.new_password as string;
-  const newPasswordConfirm = values.new_password_confirm as string;
-
-  if (newPassword.length < 8) {
-    return {
-      valid: false,
-      error: "New password must be at least 8 characters",
-    };
-  }
-  if (newPassword !== newPasswordConfirm) {
-    return { valid: false, error: "New passwords do not match" };
-  }
-
-  return { valid: true, currentPassword, newPassword };
-};
-
-/**
- * Handle POST /admin/settings
- */
-const handleAdminSettingsPost = (request: Request): Promise<Response> =>
-  withAuthForm(request, async (session, form) => {
-    const stripeKeyConfigured = await hasStripeKey();
-    const settingsPageWithError = (error: string, status: number) =>
-      htmlResponse(
-        adminSettingsPage(session.csrfToken, stripeKeyConfigured, error),
-        status,
-      );
-
-    const validation = validateChangePasswordForm(form);
-    if (!validation.valid) {
-      return settingsPageWithError(validation.error, 400);
-    }
-
-    const isCurrentValid = await verifyAdminPassword(
-      validation.currentPassword,
-    );
-    if (!isCurrentValid) {
-      return settingsPageWithError("Current password is incorrect", 401);
-    }
-
-    await updateAdminPassword(validation.newPassword);
-    return redirect("/admin/", clearSessionCookie);
-  });
-
-/**
- * Handle POST /admin/settings/stripe
- */
-const handleAdminStripePost = async (request: Request): Promise<Response> => {
-  const stripeErrorHandler = async (
-    session: AuthSession,
-    _: URLSearchParams,
-    error: string,
-  ) => {
-    const stripeKeyConfigured = await hasStripeKey();
-    return htmlResponse(
-      adminSettingsPage(session.csrfToken, stripeKeyConfigured, error),
-      400,
-    );
-  };
-
-  const result = await requireAuthValidation(
-    request,
-    stripeKeyFields,
-    stripeErrorHandler,
-  );
-  if (!result.ok) return result.response;
-
-  await updateStripeKey(result.validation.values.stripe_secret_key as string);
-  const stripeKeyConfigured = await hasStripeKey();
-  return htmlResponse(
-    adminSettingsPage(
-      result.session.csrfToken,
-      stripeKeyConfigured,
-      undefined,
-      "Stripe key updated successfully",
-    ),
-  );
-};
-
-/**
- * Handle GET /admin/sessions
- */
-const handleAdminSessionsGet = (request: Request): Promise<Response> =>
-  requireSessionOr(request, async (session) => {
-    const sessions = await getAllSessions();
-    return htmlResponse(
-      adminSessionsPage(sessions, session.token, session.csrfToken),
-    );
-  });
-
-/**
- * Handle POST /admin/sessions (log out of all other sessions)
- */
-const handleAdminSessionsPost = (request: Request): Promise<Response> =>
-  withAuthForm(request, async (session) => {
-    await deleteOtherSessions(session.token);
-    const sessions = await getAllSessions();
-    return htmlResponse(
-      adminSessionsPage(
-        sessions,
-        session.token,
-        session.csrfToken,
-        "Logged out of all other sessions",
-      ),
-    );
-  });
-
-/**
- * Handle POST /admin/event (create event)
- */
-const handleCreateEvent = createHandler(eventsResource, {
-  onSuccess: () => redirect("/admin/"),
-  onError: () => redirect("/admin/"),
-});
-
-/**
- * Handle GET /admin/event/:id
- */
-const handleAdminEventGet = (request: Request, eventId: number) =>
-  withEventAttendees(request, eventId, (event, attendees) =>
-    htmlResponse(adminEventPage(event, attendees)),
-  );
-
-/** Curried event page GET handler: renderPage -> (request, eventId) -> Response */
-const withEventPage =
-  (renderPage: (event: EventWithCount, csrfToken: string) => string) =>
-  (request: Request, eventId: number): Promise<Response> =>
-    requireSessionOr(request, (session) =>
-      withEvent(eventId, (event) =>
-        htmlResponse(renderPage(event, session.csrfToken)),
-      ),
-    );
-
-/** Render event error page or 404 */
-const eventErrorPage = async (
-  id: number,
-  renderPage: (
-    event: EventWithCount,
-    csrfToken: string,
-    error: string,
-  ) => string,
-  csrfToken: string,
-  error: string,
-): Promise<Response> => {
-  const event = await getEventWithCount(id);
-  return event
-    ? htmlResponse(renderPage(event, csrfToken, error), 400)
-    : notFoundResponse();
-};
-
-/** Handle GET /admin/event/:id/edit */
-const handleAdminEventEditGet = withEventPage(adminEventEditPage);
-
-/** Handle POST /admin/event/:id/edit */
-const handleAdminEventEditPost = updateHandler(eventsResource, {
-  onSuccess: (row) => redirect(`/admin/event/${row.id}`),
-  onError: (id, error, session) =>
-    eventErrorPage(id as number, adminEventEditPage, session.csrfToken, error),
-  onNotFound: notFoundResponse,
-});
-
-/**
- * Handle GET /admin/event/:id/export (CSV export)
- */
-const handleAdminEventExport = (request: Request, eventId: number) =>
-  withEventAttendees(request, eventId, (event, attendees) => {
-    const csv = generateAttendeesCsv(attendees);
-    const filename = `${event.name.replace(/[^a-zA-Z0-9]/g, "_")}_attendees.csv`;
-    return new Response(csv, {
-      headers: {
-        "content-type": "text/csv; charset=utf-8",
-        "content-disposition": `attachment; filename="${filename}"`,
-      },
-    });
-  });
-
-/** Handle GET /admin/event/:id/delete (show confirmation page) */
-const handleAdminEventDeleteGet = withEventPage(adminDeleteEventPage);
-
-/** Handle DELETE /admin/event/:id (delete event, optionally verify name) */
-const handleAdminEventDelete = deleteHandler(eventsResource, {
-  onSuccess: () => redirect("/admin/"),
-  onVerifyFailed: (id, _row, session) =>
-    eventErrorPage(
-      id as number,
-      adminDeleteEventPage,
-      session.csrfToken,
-      "Event name does not match. Please type the exact name to confirm deletion.",
-    ),
-  onNotFound: notFoundResponse,
-});
-
-/** Attendee with event data */
-type AttendeeWithEvent = { attendee: Attendee; event: EventWithCount };
-
-/** Load attendee ensuring it belongs to the specified event */
-const loadAttendeeForEvent = async (
-  eventId: number,
-  attendeeId: number,
-): Promise<AttendeeWithEvent | null> => {
-  const event = await getEventWithCount(eventId);
-  if (!event) return null;
-
-  const attendee = await getAttendee(attendeeId);
-  if (!attendee || attendee.event_id !== eventId) return null;
-
-  return { attendee, event };
-};
-
-/** Curried helper: load attendee for event, return 404 or apply handler */
-const withAttendeeForEvent =
-  (eventId: number, attendeeId: number) =>
-  async (
-    handler: (data: AttendeeWithEvent) => Response | Promise<Response>,
-  ): Promise<Response> => {
-    const data = await loadAttendeeForEvent(eventId, attendeeId);
-    return data ? handler(data) : notFoundResponse();
-  };
-
-/** Handle GET /admin/event/:eventId/attendee/:attendeeId/delete */
-const handleAdminAttendeeDeleteGet = async (
-  request: Request,
-  eventId: number,
-  attendeeId: number,
-): Promise<Response> =>
-  requireSessionOr(request, (session) =>
-    withAttendeeForEvent(
-      eventId,
-      attendeeId,
-    )((data) =>
-      htmlResponse(
-        adminDeleteAttendeePage(data.event, data.attendee, session.csrfToken),
-      ),
-    ),
-  );
-
-/** Verify name matches for deletion confirmation (case-insensitive, trimmed) */
-const verifyName = (expected: string, provided: string): boolean =>
-  expected.trim().toLowerCase() === provided.trim().toLowerCase();
-
-/** Handle POST /admin/event/:eventId/attendee/:attendeeId/delete with name verification */
-const handleAdminAttendeeDeletePost = async (
-  request: Request,
-  eventId: number,
-  attendeeId: number,
-): Promise<Response> =>
-  withAuthForm(request, (session, form) =>
-    withAttendeeForEvent(
-      eventId,
-      attendeeId,
-    )(async (data) => {
-      const confirmName = form.get("confirm_name") ?? "";
-      if (!verifyName(data.attendee.name, confirmName)) {
-        return htmlResponse(
-          adminDeleteAttendeePage(
-            data.event,
-            data.attendee,
-            session.csrfToken,
-            "Attendee name does not match. Please type the exact name to confirm deletion.",
-          ),
-          400,
-        );
-      }
-
-      await deleteAttendee(attendeeId);
-      return redirect(`/admin/event/${eventId}`);
-    }),
-  );
-
-// ============================================================================
-// Declarative Route Definitions
-// ============================================================================
-
-/** Parse event ID from params */
-const parseEventId = (params: RouteParams): number =>
-  Number.parseInt(params.id ?? "0", 10);
-
-/** Parse event and attendee IDs from params */
-const parseAttendeeIds = (
-  params: RouteParams,
-): { eventId: number; attendeeId: number } => ({
-  eventId: Number.parseInt(params.eventId ?? "0", 10),
-  attendeeId: Number.parseInt(params.attendeeId ?? "0", 10),
-});
-
-/** Route handler for POST/DELETE attendee delete */
-const attendeeDeleteHandler: RouteHandlerFn = (request, params) => {
-  const ids = parseAttendeeIds(params);
-  return handleAdminAttendeeDeletePost(request, ids.eventId, ids.attendeeId);
-};
-
-/** Admin routes definition */
-const adminRoutes = defineRoutes({
-  // Dashboard (pattern handles both /admin/ and /admin)
-  "GET /admin/": (request) => handleAdminGet(request),
-
-  // Authentication
-  "POST /admin/login": (request, _, server) =>
-    handleAdminLogin(request, server),
-  "GET /admin/logout": (request) => handleAdminLogout(request),
-
-  // Settings
-  "GET /admin/settings": (request) => handleAdminSettingsGet(request),
-  "POST /admin/settings": (request) => handleAdminSettingsPost(request),
-  "POST /admin/settings/stripe": (request) => handleAdminStripePost(request),
-
-  // Sessions
-  "GET /admin/sessions": (request) => handleAdminSessionsGet(request),
-  "POST /admin/sessions": (request) => handleAdminSessionsPost(request),
-
-  // Event CRUD
-  "POST /admin/event": (request) => handleCreateEvent(request),
-  "GET /admin/event/:id": (request, params) =>
-    handleAdminEventGet(request, parseEventId(params)),
-  "GET /admin/event/:id/edit": (request, params) =>
-    handleAdminEventEditGet(request, parseEventId(params)),
-  "POST /admin/event/:id/edit": (request, params) =>
-    handleAdminEventEditPost(request, parseEventId(params)),
-  "GET /admin/event/:id/export": (request, params) =>
-    handleAdminEventExport(request, parseEventId(params)),
-  "GET /admin/event/:id/delete": (request, params) =>
-    handleAdminEventDeleteGet(request, parseEventId(params)),
-  "POST /admin/event/:id/delete": (request, params) =>
-    handleAdminEventDelete(request, parseEventId(params)),
-  "DELETE /admin/event/:id/delete": (request, params) =>
-    handleAdminEventDelete(request, parseEventId(params)),
-
-  // Attendee delete
-  "GET /admin/event/:eventId/attendee/:attendeeId/delete": (
-    request,
-    params,
-  ) => {
-    const ids = parseAttendeeIds(params);
-    return handleAdminAttendeeDeleteGet(request, ids.eventId, ids.attendeeId);
-  },
-  "POST /admin/event/:eventId/attendee/:attendeeId/delete":
-    attendeeDeleteHandler,
-  "DELETE /admin/event/:eventId/attendee/:attendeeId/delete":
-    attendeeDeleteHandler,
-});
-
-/** Route admin requests using declarative router */
-export const routeAdmin = createRouter(adminRoutes);
