+++ src/routes/admin/auth.ts
@@ -1,0 +1,89 @@
+/**
+ * Admin authentication routes - login and logout
+ */
+
+import {
+  clearLoginAttempts,
+  createSession,
+  deleteSession,
+  isLoginRateLimited,
+  recordFailedLogin,
+  verifyAdminPassword,
+} from "#lib/db";
+import { validateForm } from "#lib/forms.tsx";
+import { loginFields } from "#templates";
+import { defineRoutes } from "../router.ts";
+import type { ServerContext } from "../types.ts";
+import {
+  generateSecureToken,
+  getClientIp,
+  parseFormData,
+  redirect,
+  withSession,
+} from "../utils.ts";
+import { loginResponse } from "./dashboard.ts";
+import { clearSessionCookie } from "./utils.ts";
+
+/**
+ * Handle POST /admin/login
+ */
+const handleAdminLogin = async (
+  request: Request,
+  server?: ServerContext,
+): Promise<Response> => {
+  const clientIp = getClientIp(request, server);
+
+  // Check rate limiting
+  if (await isLoginRateLimited(clientIp)) {
+    return loginResponse(
+      "Too many login attempts. Please try again later.",
+      429,
+    );
+  }
+
+  const form = await parseFormData(request);
+  const validation = validateForm(form, loginFields);
+
+  if (!validation.valid) {
+    return loginResponse(validation.error, 400);
+  }
+
+  const valid = await verifyAdminPassword(validation.values.password as string);
+  if (!valid) {
+    await recordFailedLogin(clientIp);
+    return loginResponse("Invalid credentials", 401);
+  }
+
+  // Clear failed attempts on successful login
+  await clearLoginAttempts(clientIp);
+
+  const token = generateSecureToken();
+  const csrfToken = generateSecureToken();
+  const expires = Date.now() + 24 * 60 * 60 * 1000; // 24 hours
+  await createSession(token, csrfToken, expires);
+
+  return redirect(
+    "/admin/",
+    `__Host-session=${token}; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=86400`,
+  );
+};
+
+/**
+ * Handle GET /admin/logout
+ */
+const handleAdminLogout = (request: Request): Promise<Response> =>
+  withSession(
+    request,
+    async (session) => {
+      await deleteSession(session.token);
+      return redirect("/admin/", clearSessionCookie);
+    },
+    () => redirect("/admin/", clearSessionCookie),
+  );
+
+/** Authentication routes */
+export const authRoutes = defineRoutes({
+  "POST /admin/login": (request, _, server) =>
+    handleAdminLogin(request, server),
+  "GET /admin/logout": (request) => handleAdminLogout(request),
+});
